requirements.txt
Flask==2.3.2
nltk==3.8.1
python-docx==0.8.11
